blog_project/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
psycopg2-binary==2.9.2
flake8==7.1.1
gunicorn==23.0.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0